COM-1925 - Add mixin in brute force
